feat(world): implement signed vertical world with -64..319 height (Plan 08)
- WorldHeight (min_y=-64, height=384) with checked helpers
- Sparse Chunk sections (Vec<Option<ChunkSection>>), empty sections = None
- SectionKey.section_y changed to i8, heightmap changed to i16 with NO_HEIGHT sentinel
- All chunk methods (get/set_block_local, light, fluid, state) accept wy: i32
- All dimension-bound systems (physics, void damage, mob spawn, portal, fluid,
  lighting, world mutation) use WorldHeight::contains_y / dimension.height()
- Save format data_version=2 with old 0..255 migration
- Protocol v10: ChunkData carries min_section_y + section_count
- 557 tests passing, cargo fmt/check clean

+++ src/chunk_schedule.rs
@@ -19,7 +19,7 @@ pub struct DirtySectionWork {
 #[derive(Default)]
 pub struct SectionMeshScheduler {
     pending: HashMap<SectionKey, DirtySectionWork>,
-    priority: BTreeSet<(u64, i32, u16, i32)>,
+    priority: BTreeSet<(u64, i32, i8, i32)>,
     pub in_flight: HashMap<SectionKey, SectionIdentity>,
 }
